TST: Relocate PET `srtm` simulation function test to dedicated file
Relocate PET `srtm` simulation function test to a dedicated file,
following a more logical testing rationale. Also, it avoids cluttering
the PET model testing file if more testing for the function at issue is
needed.

Eventually all simulation function tests would dwell in this file.
